Solcast.Response — Type

struct Response

Represents the response received from the Solcast API.

Fields

status_code::Int64: The HTTP status code.

url::String: The URL of the response.

data::Union{Vector{UInt8}, Nothing}: The response data as bytes or Nothing if not available.

success::Bool: true if the request was successful, false otherwise.

exception::Union{String, Nothing}: An exception message if the request failed, Nothing otherwise.

Solcast.to_dict — Function

to_dict(response::Response)

Converts the response data to a dictionary (JSON) if the response is successful.

Arguments

response::Response: The HTTP response object.

Returns

parsed::Dict{String, Any}: The parsed JSON data.

Throws

invalid_response_error: If the response is not successful.

Solcast.to_dataframe — Function

to_dataframe(response::Response)

Converts the response data to a DataFrame if the response is successful.

Arguments

response::Response: The HTTP response object.

Returns

dfs::DataFrame: The DataFrame containing the response data.

Throws

invalid_response_error: If the response is not successful.